Remove the module map
I did this because I want to test what happens when we try and include
this as a dependency into pullwalla-macos because right now it is
getting a redefinition error around libcmark between an auto generated
modulemap and this modulemap.

So my theory is that maybe it is auto generating the modulemap and
therefore we don't need the explicit one anymore or something? At least
it will give me more info to decide what to dig into further.